+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.0.4 (01/13/2018)
+
+- fixes bug where zero width strings that were either denoted using back tick or apostrophe
+- parses nested functions and colorizes those
+  - an example of a nested function would be `t.pipe(d.pipe(b.pipe()));`
+- adds console as a keyword
+- adds process as a keyword
+- adds ! as an operator
+- adds < as an operator
+- adds > as an operator
+- adds % as an operator
+- adds * as an operator
+- adds ~ as an operator
+
 # 0.0.3 (01/12/2018)
 
 - string should be interpreted first as to deal with nested entities